moved form properties create/edit actions to plugin
